role_permissions & roles

+++ server/database/migrations/0002_01_01_000004_create_users_table.php
@@ -25,7 +25,7 @@ return new class extends Migration
             $table->rememberToken();
             $table->timestamps();
 
-            $table->unsignedBigInteger('role_id')->nullable();
+            $table->unsignedInteger('role_id')->nullable();
             $table->foreign('role_id')->references('id')->on('roles')->onDelete('set null');
         });
